added update lock detection

+++ app/Http/Controllers/MyJrController.php
@@ -114,6 +114,9 @@ class MyJrController extends Controller
 
     public function edit($slug){
         $jr = $this->jrService->findBySlug($slug);
+        if($jr->is_locked == 1){
+            abort(503,'This transaction is already locked from editing');
+        }
         return view('ppu.jr_my.edit')->with([
             'jr' => $jr,
         ]);
@@ -159,6 +162,9 @@ class MyJrController extends Controller
 
     public function destroy($slug){
         $jr = $this->jrService->findBySlug($slug);
+        if($jr->is_locked == 1){
+            abort(503,'This transaction is already locked');
+        }
         if($jr->delete()){
             $jr->transDetails()->delete();
             return 1;
